Fix moving the camera when it would get clamped down
This commit renames some variables to more evident names and adds
comments.

`_adjust_camera` renamed to `_clamp_camera`, because that's a more
obvious name. It has been altered to also return whether or not the
camera was clamped, which is required for moving the camera from
close to a corner.