Clamp action menu within viewport
This alters the action menu so it has to inherit from a Container class.

The reason behind it is that Node2D boundary calculation may or may not
be impossible, but Containers do it naturally.

It has also been said that you should always use Control nodes for UI,
so this feels like a more idiomatic approach.