Fix force_text_ids behavior
Before this, enforcing the identifiers meant they were always shown
in the game. Now they are actually enforced by reporting errors.

Strings with identifiers that have not been translated do get a
complaint that's visible in-game, because that may be a work-in-progress
situation and not an error.

Also got rid of unnecessary `tid` variable.