Move `dialog_default` out of demo as a `Container`
Positioning is easier for a Container, which should be used for
UI elements anyway. This also gets rid of setting the z index.

The reason for not having it in the demo is that it's not really
a demo asset but a general thing in the framework, which is used
as a default, as the name implies.