Vertical clamp of tooltip
Horizontal is too hard because of unknown widths. Anything I tried
to do about it was wrong in relation to the pointer, but this
alone helps a lot.

Also change variable naming to use double quotes, as was discussed
on IRC. The reason we want double quotes is consistency, as $/foo/bar
will not work anyway.