ESC function to toggle HUD visibility
If your inventory is overlaid onto the game, and you want to
do camera-movement trickery, it's hard to set the scene so
the inventory doesn't get in the way.

It's also confusing for the brain because having an interactive
element feels like the game would be interactive during the
camera movement.