Do not clear params_queue until walk_stop()
There were cases where the params_queue was cleared, with the side
effect that nothing happened until opening the action menu again.

@fleskesvor suggested this has been a problem with the verb menu
as well.

My investigations lead me to believe it has to do with disjointed
terrains or other unwalkable destinations that caused the queue
reset to happen too soon.