Drop esc_type.BACKGROUND for esc_type.BACKGROUND_AREA
As can be seen in the removed documentation, using `TextureRect`
for the background caused problems that are far worse than having
to un-center a `Sprite`. These include click masks and not being
able to use parallax.

Beside that, the old background.gd file extended `Control`, which
could lead to situations where something completely bizarre was
attempted for the background and break all the things.

Now there is only one very clear-cut and documented way of doing
things.