Make player's set_state actually set a state, not just a costume
This aligns the behavior with NPC behavior without breaking
how simple items can still be altered with `set_state`. The difference
lies in how idle animations are handled and when the costume
change comes into effect.

Therefore a character's costume change must be a combination of
`cut_scene` to allow the costume change animation to finish and then
an explicit `set_state char idle` or whatever you want your character to
do next.