Implement disabling input through telon
`cut_scene telon disable_input` and `cut_scene telon enable_input`
which call a new function in telon.gd.

Use this for cutscenes which have any sort of timing
related to them, because user input like skipping dialog
would break it.

Disables saving as well, because loading a game from the
middle of a cutscene can't lead to anything good.