Settings for HiDPI/Retina
Although the documentation for
[Multiple Settings](http://docs.godotengine.org/en/3.0/tutorials/viewports/multiple_resolutions.html)
makes it sound like the `Expand` aspect ratio doesn't do
what we want, godotengine/godot#19912 states that this is the correct way.

Further documentation may be warranted for the scene tree to
avoid having a clear-color bar at the top of a Retina game.

This commit does not yet address that documentation.