Option for items to set dynamic/static z-index
This de-couples z-indexing from scaling. The use case is when
gfx artists draws a room with different "element items" as layers
in the same scale, and using the sizemap to get the z-index would
shrink the item.

Before this commit, the "scale from map" feature was responsible
for *both* the scaling *and* the z-index.