Optimization of how and when z-index is calculated
This dropped the timedelta from 0.06 to 0.02, and 0.04 when
an npc and the player move at the same time.

It initializes the player's and `Node2D` items' states in `_ready()`,
so they don't need constant updating on every tick.

The z-indexes, and thus also light and what you have in
`_update_terrain()`, are then updated only when moving - when they really
need to be.