Alleviate race condition between mouse_enter and mouse_exit
Sometimes it would happen when moving the cursor over items that
the new `mouse_entered` signal would be emitted before the
previous `mouse_exited`.

In this case the tooltip would sort of be set correctly, but then
cleared by the exit signal.

Now we yield on either the item or its area, to support both
inert `Area2D` items and `Sprite` items with an `Area2D` child.