Implement interact_status to track walking for interaction
Apparently Godot considers a click first, making the player walk
to the where an item was clicked, then resolves the interact
position, then another click executes the default action, but does
not move the player to interact position.

This would look silly even if it worked that, so the more elegant
solution is to make the player ignore the initial click position.