Make it possible to use `run_event` as a coroutine with signals
This makes it possible to eg. hide the tooltip during events.

Simply hiding the tooltip in the beginning of a function and restoring
at the end will cause it to suddenly become visible, but blocking
on a coroutine behaves as expected.

Because `mouse_entered` is emitted with a unique flag, somehow the
Unique Group Call queue is stuck in a lock unless a signal resets
the tooltip.