Fix bug where wrong node is looked for click
If you have an NPC with an `Area2D`, it's the area that gets
evaluated for being clicked, while it's the parent node that
actually receives it.

Therefore, look at the parent for clicks when the clicked area
is not an item. It would be an item for something static in the
background, but not NPCs.