Stop player on telekinetic double-click events
Triggering a telekinetic default action by double-clicking will set
the player walking, so stop as immediately as possible.

It would be better if Godot waited a while to see if the first click of
a double-click is a click or should it be handled as a double-click.

This does not look perfect, but it sure looks better than having the
player start saying something and walk over. Especially when choosing
the default action from the menu will not walk.