Implement ! as a way of ending dialog
The purpose of this is to have a clean way of hiding the HUD during
dialog. The VM gets `in_dialog` in its context and hides the HUD.

When the `!` (alias `end_dialog`) command is called, the context
is reset. The HUD is also restored unless the event flag `NO_HUD`
is set.