Deprecate background.gd's item_was_clicked()
This was not a very reliable way of checkings things, as opposed to
checking `hover_stack` for actual items.

Items without `is_interactive` never get put on the stack so removing
this had no side-effects on them.

Also drop `is_clicked()`, because it was used only in one place, and
`clicked` doesn't require a setter and getter to complete the set.