shadow_caster, the area child of light where shadows are active
This is the 1st commit message:

shadow_caster, the area child of light where shadows are active

This is the commit message #2:

Shadows are started by shadow_caster

This is the commit message #3:

Check if the shadow is inside the caster before showing

Currently we check if `player` is in `shadow_caster`.

This will cause problems when determining the opacity of the shadow
based on distance from the edge, because `shadow` can move to the
outside of `shadow_caster` - and even gain a long distance - while
`player`'s `KinematicBody2D` is still in `shadow_caster`.