Enable shadow_caster always affecting the entered body
This means you can have a different player light mask than
the light's range cull mask, but still cast a shadow and
illuminate the player when walking into the light.

The use case for this is the ability to stand in front of a
beam of light without being lit up.