Replace set_use_action_menu with set_interactive
StraToN is on record as disliking the name for `set_use_action_menu`
and it felt a bit off sometimes that something's disabled but the
tooltip would still be visible.

With this change, the question about action menu becomes moot because
`set_interactive obj false` will deactivate `obj.area` and thus make
it non-interactive.

If someone actually prefers the original behavior, it's easy to
re-implement, but this is most certainly the better way.