Do not override the character in params for dialog
This caused random crashes, which are near impossible to figure out,
where something (eventually `check_obj` in vm_level.gd) checking
for `say` is expecting a string but got an object.

As this did not happen consistently and the cost of looking up the
`say`ing object isn't super expensive, revert this part of the change.