Implement and document speed for walk/slide commands
This was a weird stub in the code, now it's a real thing.

It works by overriding the player's or item's speed setting
for the duration of the walk/slide and restoring it when
stopped.

Changed speed to be an integer so we don't accidentally have
near-zero floats being mis-evaluated as whether or not to move
at that speed.