Replace telon's input-catching with acceptable inputs
This fixes the bug where the tooltip (following the mouse),
when moved during dialog, would show the old tooltip text.

The reason for that was how telon blocked not just click events
but also the item hover events.

Because we actually care mainly about skipping dialog, make it
a state and use the main input handler to deal with it.