Fix corruption when events trigger new events
This takes the new event's code and adds it to the end
of the current code.

It also fixes an issue that would arise from only checking a running
event, namely that it's not `:setup` running after a `change_scene`
command, but the previous eg. `:hand` (on a door).

The event itself is not relevant at this point, as its code (like
`set_global last_exit outside`) has been pushed on the stack and will be
executed in due order either way.