Fix crash issue with always-open inventory and in-game menu, exiting item
This is mainly a workaround for how horrible the mouse event interaction
between `Control` and `Node2D` is.

Even if we had a way to not use `Control` in inventory, exiting a
hovered item while the in-game menu is open would still break the
tooltip when following the mouse. Unless there's a way to make all
inventory `Node2D` and pass the in-game menu's mouse events to it
so the relevant events are seen.