Use event actions instead of mouse buttons
The ones defined are:

  * game_general
  * game_rmb
  * game_highlight

One could argue there's no point in having actions if one is called
"rmb" for "right mouse button", but one could also argue that the rmb
is configurable. Games in Escoria are quite flexible too, so the action
name can't be so specific it will tie things down in different games.