Proper restoration of tooltips after in-game menu is closed
Sadly this also required adding a special case for inventory items
and disabling all sanity checking on `hover_stack`.

The HUD button refactor should also be a separate commit, but it's not
because this should also be easy to cherry-pick into the 3.1 branch...

Likewise the special case for inventory should be a separate commit,
but squashed it here for cherry-picking purposes.

And maybe the commit about the action menu as well...