Fix bug where could not transition to ESCRoom with esc_script not set.
When creating a new `ESCRoom`, there was a bug where it was not possible
to transition into the room if the room's `esc_script` was not set.
Apparently, `_perform_script_events()` would bail out early when
`room.esc_script` is falsy, which seems to be why the transition does
not happen.

Perhaps this defensive check was necessary at one time, but that does
not appear to be the case today because the only code that reads
`room.esc_script` in this file is within the `_run_script_event()`
function, which does its own defensive check for `room.esc_script`. As
such, it appears that we can let the rest of `_perform_script_events()`
do its thing even when `room.esc_script` is falsy (i.e., `""`).