feat: enable ESCItem to use custom ESCEvents written in GDScript
My high-level goal is to use my own scripting language for Escoria.
This goes back to some of my original work on RoomScript:

https://github.com/godot-escoria/escoria-demo-game/pull/558

as well as this prototype to generate ESC from my own scripting language:

https://github.com/bolinfest/alt-esc-prototype/

Because #558 was a large change that was impractical to upstream,
I decided to try to rework things as a much smaller diff so that
I could unblock myself without having to maintain a fork.
The crux of this commit is only a two-line change to `esc_object_manager.gd`.

The rest of the changes in the commit are updates to the demo to illustrate how
the feature could be used, as it changes `r3_l_exit` to be a custom instance of
`ESCItem` that implements the new, optional `generate_events()` function so it
can return its own `Dictionary` to map event names to `ESCEvent` instances.

Though in practice, I would not expect to author this sort of code by hand.
Instead, I would have my own scripting language from which I would generate
files such as `l_door.gd` and `l_door_exit_scene.gd`.

Though as illustrated by `_compile_event_from_esc()`, I could
start by generating ESC instead of going straight to GDScript.
For example, my input language might be:

```
room room03 {
  item l_door {
    event exit_scene {
      // ESC
      change_scene "res://game/rooms/room02/room02.tscn"
    }
  }

  item r_door {
    event exit_scene {
      // ESC
      change_scene "res://game/rooms/room04/room04.tscn"
    }
  }
}
```

And from this, it would be fairly straightforward to generate these two `.gd` files:

```gdscript
# game/rooms/room03/l_door.gd
extends ESCItem

func generate_events() -> Dictionary:
    return {
        "exit_scene": escoria.esc_compiler.compile([
            ":exit_scene",
            "change_scene \"res://game/rooms/room02/room02.tscn\"",
        ]).events["exit_scene"],
    }
```

and:

```gdscript
# game/rooms/room03/r_door.gd
extends ESCItem

func generate_events() -> Dictionary:
    return {
        "exit_scene": escoria.esc_compiler.compile([
            ":exit_scene",
            "change_scene \"res://game/rooms/room04/room04.tscn\"",
        ]).events["exit_scene"],
    }
```

Initially, I admit this does not buy much other than making it possible
to define events for multiple items in one file. Though over time, I
would like to be able to more sophisticated things like you see in:

https://github.com/grumpygamer/DeloresDev/blob/master/Scripts/Rooms/MainStreet.dinky

While I know it is possible to extend ESC by creating my own instances of
`ESCCommand`, I would prefer to have the freedom to experiment with a
broarder array of constructs.